Added CustomKeyboardHelper
