+++ app/src/main/res/layout/activity_login.xml
@@ -7,7 +7,6 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_gravity="center"
-        android:layout_marginBottom="@dimen/login_margin_bottom"
         android:orientation="vertical">
 
         <TextView
@@ -52,17 +51,5 @@
             android:textSize="@dimen/submit_password_text_size" />
     </LinearLayout>
 
-    <android.inputmethodservice.KeyboardView
-        android:id="@+id/keyboardview"
-        android:layout_width="fill_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true"
-        android:focusable="true"
-        android:focusableInTouchMode="true"
-        android:visibility="gone"
-        android:keyBackground="@drawable/key_background"
-        android:keyTextColor="#eee"
-        android:background="@color/color_primary"
-        android:keyTextSize="30sp"/>
+    <include layout="@layout/custom_keyboard" />
 </RelativeLayout>
