+++ app/src/main/res/layout/custom_keyboard.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<android.inputmethodservice.KeyboardView
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/keyboard_view"
+    android:layout_width="fill_parent"
+    android:layout_height="wrap_content"
+    android:layout_alignParentBottom="true"
+    android:layout_centerHorizontal="true"
+    android:focusable="true"
+    android:focusableInTouchMode="true"
+    android:visibility="gone"
+    android:keyBackground="@drawable/key_background"
+    android:keyTextColor="#eee"
+    android:background="@color/color_primary"
+    android:keyTextSize="30sp" />
